requirements.txt
langchain
langchain-openai
langchain-community
langgraph
openai
PyGithub
python-dotenv
chromadb
pytest
huggingface-hub
gradio